api/requirements.txt
# Runtime dependencies for the FastAPI service container.
fastapi==0.110.2
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
# Pin bcrypt to <4 because bcrypt 4+ raises on >72-byte secrets, which passlib 1.7.x
# hits during backend detection, triggering "password cannot be longer than 72 bytes".
bcrypt<4
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
httpx==0.27.0
tenacity==8.3.0
python-slugify==8.0.4
email-validator==2.1.1
python-dotenv==1.0.1
redis>=4.5,<5
rq==1.2.0
rq-scheduler==0.13.1
